Huge reorganization resulting in move to Jekyll.  The website is now built with Jekyll, which you can run via the ./create_website script.
The repo now includes the full stageoneeducation.com website, not just the rocket workshop.  This was basically needed to make everything sane for Jekyll.  There's still some issues where we duplicate basically all of bootstrap in the folders, but we can clean that up later if we want.

Now the process to build is:

./create_website   <-- this will create the www folder.  That folder is *NOT TO BE MODIFIED*. To modify the site, change files in the jekyll folder!
gcloud app deploy  <-- deploys the site to google cloud services.

Note: the www folder is now *generated* and thus is ignored by git.